move {Close, EncWriter, DecWriter} to separate module
This commit moves the `Close` trait, `EncWriter` and
`DecWriter` to the `writer` module. I do this to
prepare for the `EncReader` and `DecReader` implementation.
Keeping everything in lib.rs is not readable.
